Refactor: Factored out Cli class from Ui
To facilitate creation of the new GUI, factored out Cli as an
implementation of the Ui interface, which the new GUI component
should implement.
